Add Header and PostForm component to Home screen
Home screen is supposed to be composed of two sections; header and main element.

The header contains the PostForm component.

The main section contains the posts list.

An Avatar UI element is added that is used is reusable and cusomizable, it could take an image uri to render, otherwise it would render an anynomous avatar image, it also take size prop to customiz size to fit reply or any where else.